+++ api/test/unit/domain/analyzers/test_integrated_alerts_analyzer.py
@@ -7,10 +7,13 @@ import pytest
 import rioxarray  # noqa: F401
 import xarray as xr
 
+from app.domain.analyzers import integrated_alerts_analyzer
 from app.domain.analyzers.integrated_alerts_analyzer import IntegratedAlertsAnalyzer
+from app.domain.models.analysis import Analysis
 from app.infrastructure.external_services.duck_db_query_service import (
     DuckDbPrecalcQueryService,
 )
+from app.models.land_change.integrated_alerts import IntegratedAlertsAnalyticsIn
 
 # Pixel areas down one column of latitude; reused for both fixtures and for
 # computing expected per-group sums so nothing is a magic number.
@@ -34,7 +37,94 @@ Y_VALS = np.linspace(48.0, 47.99775, 10)  # latitude, descending (matches clip s
 X_VALS = np.linspace(105.0, 105.00225, 10)  # longitude, ascending
 
 
-class TestIntegratedAlertsPrecomputedAdminAnalysis:
+class FakeQueryService:
+    def __init__(self):
+        self.query = None
+
+    async def execute(self, query):
+        self.query = query
+        return {
+            "aoi_id": ["IDN.24.9"],
+            "alert_date": ["2029-06-01"],
+            "alert_confidence": ["high"],
+            "area_ha": [1.0],
+        }
+
+
+def make_analysis(aoi, start_date="2029-01-01", end_date="2032-12-31") -> Analysis:
+    metadata = IntegratedAlertsAnalyticsIn(
+        aoi=aoi, start_date=start_date, end_date=end_date
+    ).model_dump()
+    return Analysis(result=None, metadata=metadata, status="pending")
+
+
+class TestAnalyzerRouting:
+    """analyze() dispatch and date normalization. The query/transform behavior
+    itself is covered by the analysis tests below."""
+
+    @pytest.mark.asyncio
+    async def test_admin_normalizes_year_only_dates(self):
+        query_service = FakeQueryService()
+        analyzer = IntegratedAlertsAnalyzer(
+            duckdb_query_service=query_service, input_uris={}
+        )
+        analysis = make_analysis(
+            {"type": "admin", "ids": ["IDN", "BRA.1"]},
+            start_date="2029",
+            end_date="2032",
+        )
+
+        await analyzer.analyze(analysis)
+
+        # year-only inputs are widened to a full date window before querying
+        assert "2029-01-01" in query_service.query
+        assert "2032-12-31" in query_service.query
+        assert analysis.result["aoi_type"] == ["admin"]
+
+    @pytest.mark.asyncio
+    async def test_admin_passes_full_dates_through(self):
+        query_service = FakeQueryService()
+        analyzer = IntegratedAlertsAnalyzer(
+            duckdb_query_service=query_service, input_uris={}
+        )
+        analysis = make_analysis(
+            {"type": "admin", "ids": ["IDN.24.9"]},
+            start_date="2029-03-16",
+            end_date="2030-06-01",
+        )
+
+        await analyzer.analyze(analysis)
+
+        assert "2029-03-16" in query_service.query
+        assert "2030-06-01" in query_service.query
+
+    @pytest.mark.asyncio
+    async def test_non_admin_routes_to_otf(self, monkeypatch):
+        class RoutedToOtf(Exception):
+            pass
+
+        class FakeEngine:
+            def map(self, *args, **kwargs):
+                raise RoutedToOtf()
+
+            async def gather(self, *args, **kwargs):
+                raise AssertionError("gather should not be reached")
+
+        async def fake_get_geojson(aois):
+            return [{"type": "Polygon", "coordinates": []}]
+
+        monkeypatch.setattr(integrated_alerts_analyzer, "get_geojson", fake_get_geojson)
+
+        analyzer = IntegratedAlertsAnalyzer(
+            compute_engine=FakeEngine(), input_uris={"x": "y"}
+        )
+        analysis = make_analysis({"type": "protected_area", "ids": ["555625448"]})
+
+        with pytest.raises(RoutedToOtf):
+            await analyzer.analyze(analysis)
+
+
+class TestPrecomputedAdminAnalysis:
     @pytest.fixture
     def precomputed_admin_results(self, tmp_path):
         # Parquet keyed by aoi_id and preaggregated to grain
@@ -94,7 +184,7 @@ class TestIntegratedAlertsPrecomputedAdminAnalysis:
         pd.testing.assert_frame_equal(expected, df, check_like=True, check_dtype=False)
 
 
-class TestIntegratedAlertsOtfAnalysis:
+class TestOtfAnalysis:
     @pytest.fixture
     def alerts_datacube(self):
         # alert_date constant; confidence in three bands so 2/3/4 -> low/high/highest
