+++ api/test/unit/models/land_change/test_integrated_alerts.py
@@ -1,0 +1,35 @@
+import pytest
+from pydantic import ValidationError
+
+from app.models.land_change.integrated_alerts import (
+    ANALYTICS_NAME,
+    IntegratedAlertsAnalyticsIn,
+)
+
+
+class TestIntegratedAlertsAnalyticsIn:
+    def test_defaults(self):
+        analytics_in = IntegratedAlertsAnalyticsIn(
+            aoi={"type": "admin", "ids": ["IDN.24.9"]},
+            start_date="2024-01-01",
+            end_date="2024-12-31",
+        )
+        assert ANALYTICS_NAME == "integrated_alerts"
+        assert analytics_in._analytics_name == "integrated_alerts"
+
+    def test_rejects_intersections_field(self):
+        with pytest.raises(ValidationError):
+            IntegratedAlertsAnalyticsIn(
+                aoi={"type": "admin", "ids": ["IDN.24.9"]},
+                start_date="2024-01-01",
+                end_date="2024-12-31",
+                intersections=["driver"],
+            )
+
+    def test_rejects_bad_date(self):
+        with pytest.raises(ValidationError):
+            IntegratedAlertsAnalyticsIn(
+                aoi={"type": "admin", "ids": ["IDN.24.9"]},
+                start_date="not-a-date",
+                end_date="2024-12-31",
+            )
